Refactor quote schemas
