+++ jaz_api_v03/src/quotes/routes.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import Session
 
 from ..core.dependencies import get_oracle_session, get_session
-from . import crud, schemas
+from . import crud, schemas, schemas_
 
 router = APIRouter()
 
@@ -15,7 +15,7 @@ router = APIRouter()
 async def send_payload(
     *,
     async_db: AsyncSession = Depends(get_session),
-    payload_in: schemas.PartnerTransBase,
+    payload_in: schemas_.PartnerTransBase,
     # current_user: models.User = Depends(deps.get_current_active_superuser),
 ) -> Any:
     # 1. Search and/or create customer details in premia
@@ -25,14 +25,21 @@ async def send_payload(
     #     cvr_sr_no=1,
     #     prc_code="3553",
     #     prc_rate=1.75,
-    #     prc_rate_per=100,
+    #     prc_rate_per=1000,
     #     prc_si_curr_code="KES",
     #     prc_prem_curr_code="KES",
     #     prc_si_fc=payload_in.items_total_cost,
     #     prc_prem_fc=payload_in.premium.basic_prem,
     # )
 
-    # 2. Create quote
+    # 3. create proposal smi
+    items = payload_in.items
+    for item in enumerate(items, 1):
+        sr_no, smi = item
+
+    # proposal_smi_obj = schemas.ProposalSMICreate(smi_sr_no=1)
+
+    # 8. Create quote
     quote_obj = schemas.QuoteCreate(
         quot_ref=payload_in.trans_ref,
         quot_paymt_ref=payload_in.prem_payment_ref,
@@ -43,7 +50,7 @@ async def send_payload(
     # 3. create proposal
     # 4. create section
     # 5. Create risk
-    # 6. create smi
+
     # 7. create cover
     # 8. create charges
 
@@ -56,7 +63,7 @@ async def send_payload(
 def test_oracle(
     *,
     oracle_db: Session = Depends(get_oracle_session),
-    payload_in: schemas.PartnerTransBase,
+    payload_in: schemas_.PartnerTransBase,
     # current_user: models.User = Depends(deps.get_current_active_superuser),
 ) -> Any:
     result = oracle_db.execute(text("select * from jick_t where rownum<=5"))
