+++ jaz_api_v03/src/quotes/schemas/__init__.py
@@ -1,0 +1,13 @@
+from .proposalcover import (  # noqa: F401
+    ProposalCover,
+    ProposalCoverCreate,
+    ProposalCoverInDB,
+    ProposalCoverUpdate,
+)
+from .proposalsmi import (  # noqa: F401
+    ProposalSMI,
+    ProposalSMICreate,
+    ProposalSMIInDB,
+    ProposalSMIUpdate,
+)
+from .quote import Quote, QuoteCreate, QuoteInDB, QuoteUpdate  # noqa: F401
